+++ app/scripts/lib/pending-tx-tracker.js
@@ -1,6 +1,5 @@
 const EventEmitter = require('events')
 const EthQuery = require('ethjs-query')
-const sufficientBalance = require('./util').sufficientBalance
 /*
 
   Utility class for tracking the transactions as they
@@ -12,7 +11,6 @@ const sufficientBalance = require('./util').sufficientBalance
   requires a: {
     provider: //,
     nonceTracker: //see nonce tracker,
-    getBalnce: //(address) a function for getting balances,
     getPendingTransactions: //() a function for getting an array of transactions,
     publishTransaction: //(rawTx) a async function for publishing raw transactions,
   }
@@ -25,7 +23,6 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
     this.query = new EthQuery(config.provider)
     this.nonceTracker = config.nonceTracker
     this.retryLimit = config.retryLimit || Infinity
-    this.getBalance = config.getBalance
     this.getPendingTransactions = config.getPendingTransactions
     this.publishTransaction = config.publishTransaction
   }
@@ -89,31 +86,22 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
         // other
         || errorMessage.includes('gateway timeout')
         || errorMessage.includes('nonce too low')
-        || txMeta.retryCount > 1
       )
       // ignore resubmit warnings, return early
       if (isKnownTx) return
       // encountered real error - transition to error state
-      this.emit('tx:failed', txMeta.id, err)
+      txMeta.warning = {
+        error: errorMessage,
+        message: 'There was an error when resubmitting this transaction.',
+      }
+      this.emit('tx:warning', txMeta, err)
     }))
   }
 
   async _resubmitTx (txMeta) {
-    const address = txMeta.txParams.from
-    const balance = this.getBalance(address)
-    if (balance === undefined) return
-
     if (txMeta.retryCount > this.retryLimit) {
       const err = new Error(`Gave up submitting after ${this.retryLimit} blocks un-mined.`)
       return this.emit('tx:failed', txMeta.id, err)
-    }
-
-    // if the value of the transaction is greater then the balance, fail.
-    if (!sufficientBalance(txMeta.txParams, balance)) {
-      const insufficientFundsError = new Error('Insufficient balance during rebroadcast.')
-      this.emit('tx:failed', txMeta.id, insufficientFundsError)
-      log.error(insufficientFundsError)
-      return
     }
 
     // Only auto-submit already-signed txs:
@@ -148,11 +136,10 @@ module.exports = class PendingTransactionTracker extends EventEmitter {
       }
     } catch (err) {
       txMeta.warning = {
-        error: err,
+        error: err.message,
         message: 'There was a problem loading this transaction.',
       }
-      this.emit('tx:warning', txMeta)
-      throw err
+      this.emit('tx:warning', txMeta, err)
     }
   }